refactor: rearrange and rename commands to
