+++ bin/aspiresync
@@ -1,15 +1,15 @@
 #!/usr/bin/env php
 <?php
 
-use AspirePress\AspireSync\Commands\Plugins\PluginsDownloadCommand;
-use AspirePress\AspireSync\Commands\Plugins\PluginsDownloadSingleCommand;
-use AspirePress\AspireSync\Commands\Plugins\PluginsMetaCommand;
-use AspirePress\AspireSync\Commands\Plugins\PluginsMetaDumpCommand;
+use AspirePress\AspireSync\Commands\Download\DownloadPluginsCommand;
+use AspirePress\AspireSync\Commands\Download\DownloadPluginsSingleCommand;
+use AspirePress\AspireSync\Commands\Meta\MetaSyncPluginsCommand;
+use AspirePress\AspireSync\Commands\Meta\MetaDumpPluginsCommand;
 use AspirePress\AspireSync\Commands\SandboxCommand;
-use AspirePress\AspireSync\Commands\Themes\ThemesDownloadCommand;
-use AspirePress\AspireSync\Commands\Themes\ThemesDownloadSingleCommand;
-use AspirePress\AspireSync\Commands\Themes\ThemesMetaCommand;
-use AspirePress\AspireSync\Commands\Themes\ThemesMetaDumpCommand;
+use AspirePress\AspireSync\Commands\Download\DownloadThemesCommand;
+use AspirePress\AspireSync\Commands\Download\DownloadThemesSingleCommand;
+use AspirePress\AspireSync\Commands\Meta\MetaSyncThemesCommand;
+use AspirePress\AspireSync\Commands\Meta\MetaDumpThemesCommand;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\Console\Application;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -32,15 +32,15 @@ $CONTAINER->compile(resolveEnvPlaceholders: true);
 
 $APP = new Application();
 
-$APP->add($CONTAINER->get(PluginsDownloadCommand::class));
-$APP->add($CONTAINER->get(PluginsDownloadSingleCommand::class));
-$APP->add($CONTAINER->get(PluginsMetaCommand::class));
-$APP->add($CONTAINER->get(PluginsMetaDumpCommand::class));
+$APP->add($CONTAINER->get(DownloadPluginsCommand::class));
+$APP->add($CONTAINER->get(DownloadPluginsSingleCommand::class));
+$APP->add($CONTAINER->get(MetaSyncPluginsCommand::class));
+$APP->add($CONTAINER->get(MetaDumpPluginsCommand::class));
 
-$APP->add($CONTAINER->get(ThemesDownloadCommand::class));
-$APP->add($CONTAINER->get(ThemesDownloadSingleCommand::class));
-$APP->add($CONTAINER->get(ThemesMetaCommand::class));
-$APP->add($CONTAINER->get(ThemesMetaDumpCommand::class));
+$APP->add($CONTAINER->get(DownloadThemesCommand::class));
+$APP->add($CONTAINER->get(DownloadThemesSingleCommand::class));
+$APP->add($CONTAINER->get(MetaSyncThemesCommand::class));
+$APP->add($CONTAINER->get(MetaDumpThemesCommand::class));
 
 $APP->add($CONTAINER->get(SandboxCommand::class));
