+++ src/Commands/Download/DownloadPluginsCommand.php
@@ -2,14 +2,13 @@
 
 declare(strict_types=1);
 
-namespace AspirePress\AspireSync\Commands\Plugins;
+namespace AspirePress\AspireSync\Commands\Download;
 
-use AspirePress\AspireSync\Commands\AbstractDownloadCommand;
 use AspirePress\AspireSync\Services\Plugins\PluginListService;
 use AspirePress\AspireSync\Services\Plugins\PluginMetadataService;
 use AspirePress\AspireSync\Services\ProcessManager;
 
-class PluginsDownloadCommand extends AbstractDownloadCommand
+class DownloadPluginsCommand extends AbstractDownloadCommand
 {
     public function __construct(
         PluginListService $listService,
